Making methods of ICaloReadCrosstalkMap const.
We want to make the methods of ICaloReadCrosstalkMap const, for thread-safety
and general cleanliness.
Here, we add change ReadCaloCrosstalkMap::getNeighbours and getCrosstalks
to be const.
However, for now (until the interface class is changed) we also
retain the non-const overloads.

Make ICaloReadCrosstalkMap::getNeighbours and getCrosstalks const.
For now (until the interface class is changed) we retain the non-const
overloads.